Show game scores
As a beer league hockey player,
When I open the Wolfpack app,
And I tap the tab titled "Games",
And I see a list of the dates and times of my team's games,
Then I also see a score by each game (my team's score - the other team's score) (in green/red for wins/losses) and a W or an L indicating win or loss,
So that I can estimate the likelihood of my team making the playoffs and having additional games.

No score is shown by games that have not been played.

Ties are shown in black.

https://www.pivotaltracker.com/story/show/187029132