RSVP to games
Allow players to RSVP to games, which also meant adding authentication.


This commit adds rails credentials to the application.

Before deploying this commit, we should add the value in the (now
git-ignored) `config/credentials/production.key` file as value for
Heroku's `RAILS_MASTER_KEY` environment variable.

Heroku is where we host the app.