Create team model (deploy part 1)
Deploy this commit and run the migrations.

Then create a single team in the production rails console for the
existing players (named "Wolfpack Thursday Upper").

This is part 1 of a several part deploy. See the following commits for
next steps.

---

Games fetched from the API need to preload their players to avoid an N+1
query and currently there is no relation between games and players
because there is implicitly one team for all games, which includes all
the existing players.

The impetus for this several part deploy is to remove that N+1
query when requesting games with their players and player attendances.

Even though this change is being made only for performance reasons at
the moment, we will also need this change for feature requirements in
the future.

---

After all of the steps that are required for this deploy are completed,
teams will have many players and teams will have many games so that
when games are loaded through the api by the frontend, games will be
able to include + preload their relationships to players and the avoid
N+1 queries which the current game serializers which load Player.all
and PlayerAttendance.all are currently causing.