Update active admin
We updated some columns and moved the ransackable methods out of each
model and into the ApplicationRecord so that we don't have to keep
copying and pasting those two methods into each model we add to Active
Admin.

We (I) really don't understand the ransackable things. Models error
without those two methods in AA, though - or if we omit an attribute or
a relation (for instance, confirmation_code) then AA crashes with some
Ransackable-originating-error.