Permit deleting players and teams
When deleting a player or team (from the rails console or through active
admin - which is the only way to delete them currently), delete their
relationship to each other. There's a join table between players and
teams and any entry in that table which joins a player to a team should
be destroyed if either the player or the team is deleted.