Fix order dependent spec failing on CI
It does not matter what order players are returned. This spec is failing
on CI because the previous assertion required the ordering of the
players returned in the json by the endpoint was the same order in which
the players were created. This commit removes the assertion requiring
any ordering of the players returned in the json.