Suppress unwanted terminal output during tests
When we send a confirmation code to a phone number as a user logs in or
signs up, instead of actually sending the text message in the
development environment, we log their confirmation code on the rails
server. We do that because sending real text messages costs money, and
while we're developing we can see the server logs. However, testing that
functionality required mocking the rails environment in the test to be
development instead of test, which caused the test to actually output
the confirmation code message when it was run. We just want to see green
dots when we run the tests, without the confirmation code message noise.