Return individual teams
The app has a screen to invite new players to a team. That screen needs
information about the team that the player is being invited to. This
app creates an endpoint for that screen to reach out and get the info
for that team.

For context, typically the add player screen won't need to make a
request to get the team because it'll already have been loaded and
cached in the app when the list of all the inviting players teams were
fetched. However, to keep deep linking working, we need this endpoint.
A user may be deep linked to the player invite screen, in which case
they'll need to make this request.