Refactor players controller routing
When we hit the /teams/:team_id/players route, we had been using the
players controller to return those players. We'd really like to be using
the TeamPlayersController because those player resources are scoped to a
team. This commit makes that change. The reasoning behind the change
made here is the same as the reasoning described in this SO post:

https://stackoverflow.com/questions/12553870/using-resources-with-custom-controller-names